refactor: rename spcp-myinfo to spcp factory
