+++ src/app/routes/admin-forms.server.routes.js
@@ -13,7 +13,7 @@ let submissions = require('../../app/controllers/submissions.server.controller')
 const emailSubmissions = require('../../app/controllers/email-submissions.server.controller')
 let encryptSubmissions = require('../../app/controllers/encrypt-submissions.server.controller')
 let PERMISSIONS = require('../utils/permission-levels').default
-const spcpFactory = require('../factories/spcp-myinfo.factory')
+const spcpFactory = require('../factories/spcp.factory')
 const webhookVerifiedContentFactory = require('../factories/webhook-verified-content.factory')
 const AdminFormController = require('../modules/form/admin-form/admin-form.controller')
 const { withUserAuthentication } = require('../modules/auth/auth.middlewares')
