+++ src/app/routes/public-forms.server.routes.js
@@ -9,7 +9,7 @@ const submissions = require('../../app/controllers/submissions.server.controller
 const encryptSubmissions = require('../../app/controllers/encrypt-submissions.server.controller')
 const emailSubmissions = require('../../app/controllers/email-submissions.server.controller')
 const { celebrate, Joi, Segments } = require('celebrate')
-const spcpFactory = require('../factories/spcp-myinfo.factory')
+const spcpFactory = require('../factories/spcp.factory')
 const webhookVerifiedContentFactory = require('../factories/webhook-verified-content.factory')
 const { CaptchaFactory } = require('../factories/captcha.factory')
 const { limitRate } = require('../utils/limit-rate')
